fix: restore sign on non-NASA Cmd20 temperatures
The Temperature refactor left the explanatory comment in place but dropped the
cast it describes:

    // Cast to int8_t first to preserve sign (uint8_t wraps negative values)
    float pipe_in_temp = command20.pipe_in.to_celsius();

Temperature stores its value in a uint8_t, so a decoded -5 C comes back as 251.
The CmdC0 outdoor path still had the cast, which is why outdoor temperatures
read correctly while the evaporator ones did not. test_cmd20_eva_temperatures
caught it once the suite started compiling again.

Cast at the four Cmd20 call sites, matching the CmdC0 path: pipe_in, pipe_out,
target_temp and room_temp. The last two are normally positive, so nothing
observed them, but they have the same defect.

The real fix is to store the value signed inside Temperature and drop these
casts. That needs care - Celsius decodes as data - 55, which exceeds int8_t for
raw values above 182 - so it is left for someone who can run the suite.

Also revert the continue-on-error on the non-NASA step. A test that cannot fail
the build is not a gate.

+++ components/samsung_ac/protocol_non_nasa.cpp
@@ -894,20 +894,20 @@ namespace esphome
                 // These are sensor readings and should always be published, regardless of pending control messages
                 // Compare to CmdC0 and Cmd8D handlers which explicitly do not check for pending control messages
                 // Cast to int8_t first to preserve sign (uint8_t wraps negative values), then to float
-                float pipe_in_temp = nonpacket_.command20.pipe_in.to_celsius();
-                float pipe_out_temp = nonpacket_.command20.pipe_out.to_celsius();
+                float pipe_in_temp = static_cast<float>(static_cast<int8_t>(nonpacket_.command20.pipe_in.to_celsius()));
+                float pipe_out_temp = static_cast<float>(static_cast<int8_t>(nonpacket_.command20.pipe_out.to_celsius()));
                 target->set_indoor_eva_in_temperature(nonpacket_.src, pipe_in_temp);
                 target->set_indoor_eva_out_temperature(nonpacket_.src, pipe_out_temp);
 
                 if (!pending_control_message)
                 {
                     last_command20s_[nonpacket_.src] = nonpacket_.command20;
-                    target->set_target_temperature(nonpacket_.src, nonpacket_.command20.target_temp.to_celsius());
+                    target->set_target_temperature(nonpacket_.src, static_cast<float>(static_cast<int8_t>(nonpacket_.command20.target_temp.to_celsius())));
                     // TODO
                     target->set_water_outlet_target(nonpacket_.src, false);
                     // TODO
                     target->set_target_water_temperature(nonpacket_.src, false);
-                    target->set_room_temperature(nonpacket_.src, nonpacket_.command20.room_temp.to_celsius());
+                    target->set_room_temperature(nonpacket_.src, static_cast<float>(static_cast<int8_t>(nonpacket_.command20.room_temp.to_celsius())));
                     target->set_power(nonpacket_.src, nonpacket_.command20.power);
                     // TODO
                     target->set_water_heater_power(nonpacket_.src, false);
